chore(annotations): move things into separate packages
