+++ cloud-annotations/src/main/java/cloud/commandframework/annotations/Permission.java
@@ -23,43 +23,23 @@
 //
 package cloud.commandframework.annotations;
 
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
 import org.checkerframework.checker.nullness.qual.NonNull;
 
 /**
- * Processor that intercepts all cloud annotation strings.
- *
- * @since 1.7.0
+ * Equivalent to {@link cloud.commandframework.Command.Builder#permission(String)}
  */
-@FunctionalInterface
-public interface StringProcessor {
+@Target({ElementType.METHOD, ElementType.TYPE})
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Permission {
 
     /**
-     * Returns a string processor that simply returns the input string.
+     * Get the command permission
      *
-     * @return no-op string processor
+     * @return Command permission
      */
-    static @NonNull StringProcessor noOp() {
-        return new NoOpStringProcessor();
-    }
-
-    /**
-     * Processes the {@code input} string and returns the processed result.
-     * <p>
-     * This should always return a non-{@code null} result. If the input string
-     * isn't applicable to the processor implementation, the original string should
-     * be returned.
-     *
-     * @param input the input string
-     * @return the processed string
-     */
-    @NonNull String processString(@NonNull String input);
-
-
-    final class NoOpStringProcessor implements StringProcessor {
-
-        @Override
-        public @NonNull String processString(final @NonNull String input) {
-            return input;
-        }
-    }
+    @NonNull String value() default "";
 }
