+++ cloud-annotations/src/main/java/cloud/commandframework/annotations/assembler/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Assemblers that creates objects from descriptors.
+ */
+package cloud.commandframework.annotations.assembler;
