+++ cloud-annotations/src/main/java/cloud/commandframework/annotations/extractor/CommandExtractor.java
@@ -21,7 +21,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 //
-package cloud.commandframework.annotations;
+package cloud.commandframework.annotations.extractor;
 
 import cloud.commandframework.annotations.descriptor.CommandDescriptor;
 import java.util.Collection;
