+++ cloud-annotations/src/main/java/cloud/commandframework/annotations/extractor/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Extractors extract information and create descriptors.
+ */
+package cloud.commandframework.annotations.extractor;
