+++ cloud-annotations/src/main/java/cloud/commandframework/annotations/string/PropertyReplacingStringProcessor.java
@@ -21,7 +21,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 //
-package cloud.commandframework.annotations;
+package cloud.commandframework.annotations.string;
 
 import java.util.function.Function;
 import java.util.regex.MatchResult;
