+++ cloud-annotations/src/main/java/cloud/commandframework/annotations/string/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * String processing utilities.
+ */
+package cloud.commandframework.annotations.string;
