+++ cloud-annotations/src/test/java/cloud/commandframework/annotations/PatternReplacingStringProcessorTest.java
@@ -23,6 +23,7 @@
 //
 package cloud.commandframework.annotations;
 
+import cloud.commandframework.annotations.string.PatternReplacingStringProcessor;
 import java.util.function.Function;
 import java.util.regex.MatchResult;
 import java.util.regex.Pattern;
