+++ cloud-annotations/src/test/java/cloud/commandframework/annotations/PropertyReplacingStringProcessorTest.java
@@ -23,6 +23,7 @@
 //
 package cloud.commandframework.annotations;
 
+import cloud.commandframework.annotations.string.PropertyReplacingStringProcessor;
 import java.util.function.Function;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
