+++ cloud-annotations/src/test/java/cloud/commandframework/annotations/feature/ArgumentWithoutAnnotationTest.java
@@ -26,10 +26,10 @@ package cloud.commandframework.annotations.feature;
 import cloud.commandframework.CommandManager;
 import cloud.commandframework.annotations.AnnotationParser;
 import cloud.commandframework.annotations.Command;
-import cloud.commandframework.annotations.ParameterNameExtractor;
-import cloud.commandframework.annotations.StandardArgumentExtractor;
 import cloud.commandframework.annotations.TestCommandManager;
 import cloud.commandframework.annotations.TestCommandSender;
+import cloud.commandframework.annotations.extractor.ParameterNameExtractor;
+import cloud.commandframework.annotations.extractor.StandardArgumentExtractor;
 import cloud.commandframework.context.CommandContext;
 import cloud.commandframework.execution.CommandResult;
 import com.google.common.base.CaseFormat;
