+++ examples/example-bukkit/src/main/java/cloud/commandframework/examples/bukkit/annotations/feature/PermissionExample.java
@@ -25,7 +25,7 @@ package cloud.commandframework.examples.bukkit.annotations.feature;
 
 import cloud.commandframework.annotations.AnnotationParser;
 import cloud.commandframework.annotations.Command;
-import cloud.commandframework.annotations.CommandPermission;
+import cloud.commandframework.annotations.Permission;
 import cloud.commandframework.examples.bukkit.ExamplePlugin;
 import cloud.commandframework.examples.bukkit.annotations.AnnotationFeature;
 import org.bukkit.Bukkit;
@@ -48,7 +48,7 @@ public final class PermissionExample implements AnnotationFeature {
         annotationParser.parse(this);
     }
 
-    @CommandPermission("example.annotation.disableplugin")
+    @Permission("example.annotation.disableplugin")
     @Command("annotations disableplugin")
     public void disablePlugin() {
         Bukkit.getServer().getPluginManager().disablePlugin(this.examplePlugin);
